Resolved changes from PR
