+++ client/src/ui/messaging/MessageCard.tsx
@@ -13,12 +13,6 @@ export interface MessageCardProps {
   scrollDown?: () => void;
 }
 
-export interface Message {
-  messageId: string;
-  message: string;
-  dateCreated?: Date;
-}
-
 
 /**
  * UI that contains the metadata of a message
